requirements.txt
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
colorama==0.4.4
et-xmlfile==1.1.0
Flask==1.1.2
Flask-Cors==3.0.10
idna==2.10
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
openpyxl==3.0.7
pandas==1.2.3
psycopg2==2.8.6
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
requests==2.24.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.25.11
Werkzeug==2.0.1
xlrd==2.0.1
zipp==3.5.0
